Added a contact_direction to Constraint-Features
this contact direction specifies from where contact can be made
with the feature most easily. The alternative, having a second
Feature for some constraints looked more clumsy.

With this addition, a feature defines a frame, so _any_ constraint
between two frames can be represented. This provides a neat way
to plug in any task functions we like.