feature_viz: removed unnecessary .v() method for Features
This method returned a LocaatedVector for a Feature. Now, compute()
does the same, getting rid of some complications in delegating transformations.

The overall goal is to have the 'functions' like Proj_P or Cos decide
which parent parts need to be shown. Thus, this complication is kept out
of the 'user' syntax.