Incorporated four bug fixes.
1.) Removed assumption that beginning of the kinematic is always at the base
link of the robot. To do this an extra transform between the robot base and
the base of the kinematic chain is included and a subscription to a respective
topic have been added.

2.) Added another subscription to a topic that publishes the pose of the
base of the robot in the world frame.

3.) ConstraintController::Update() shall only have an effect _after_ the
controller has been properly configured _and_ started.

4.) The constraint_command_callback shall only try copying the content of
the command message if it safe to do so. That means: _after_ the controller
has been configured and when the dimensions of the internal structures and
the message structures for constraints match.